AB#237353
Document component updates and bump version

Updated the Umbraco document componet to pull the title using the following rules

If a link title is set, it will be used as the title.

If no link title is set, but a document title is set in the media library, the document title will be used.

If neither a link title nor a document title is set, the document node name will be used as the title.